+++ test/specs/enum_test.dart
@@ -201,10 +201,8 @@ void main() {
 
       const MyEnum();
 
-      const factory MyEnum.redirect([
-          int? myInt,
-          String? myString,
-      ]) = MyOtherEnum.named;
+      const factory MyEnum.redirect([int? myInt, String? myString])
+          = MyOtherEnum.named;
     }
     '''));
   });
@@ -248,10 +246,7 @@ void main() {
       b(1, 'abc'),
       c;
 
-      const MyEnum([
-          this.myInt,
-          this.myString,
-      ]);
+      const MyEnum([this.myInt, this.myString]);
 
       final int? myInt;
       final String? myString;